+++ config/env.docker.sh
@@ -13,14 +13,15 @@ export ALLOW_CHANGE_EMAIL=false
 export I18N_ENABLED=true
 export ALLOW_CHANGE_PASSWORD=false
 export ALLOW_CHANGE_LANG=false
+export ALLOW_SIGNUP_PATIENT=false
 export HIDE_DONATE=true
 export HIDE_DEXCOM_BANNER=true
 export HIDE_UPLOAD_LINK=true
 export BRANDING='diabeloop'
 export PASSWORD_MIN_LENGTH=10
 export PASSWORD_MAX_LENGTH=72
-# url to HELP system 
+# url to HELP system
 export HELP_LINK='https://url.com'
-# url to web server hosting the asset files 
-# such as terms of use, data privacy, etc. 
+# url to web server hosting the asset files
+# such as terms of use, data privacy, etc.
 export ASSETS_URL='https://url.com'