+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Blip 
 
+## [Unreleased]
+
+### Fixed
+- PT-532 One can create a patient even if the application is not allowing it
+
 ## [0.5.0] - 2019-07-29
 
 ### Added